Separate 'RunR8' task
